requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.38.0
optimum[onnxruntime]>=1.17.0
onnxruntime>=1.17.0
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
fastwer>=0.1.3
streamlit>=1.31.0
scipy>=1.11.0
Pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
